refactor(summary-card): modularize and add loading state
Move category-related logic to separate files (types, utils, styled) for better maintainability. Add loading state to SummaryCard component to improve user experience during data fetching.
