+++ packages/ui/src/main/home-page/summary-card/types.ts
@@ -1,0 +1,3 @@
+type CategoryType = 'must' | 'nice' | 'waste' | 'total';
+
+export { type CategoryType };
